requirements.txt
django==4.1.2
black==23.1.0
djangorestframework==3.14.0
django-cors-headers==3.13.0